scripts/requirements.txt
bcrypt>=4.0.1
psycopg2-binary>=2.9.7